requirements.txt
numpy>=1.26
pandas>=2.2
scipy>=1.13
matplotlib>=3.9
scikit-learn>=1.5
xgboost>=2.1
statsmodels>=0.14
arch>=6.3
pyyaml>=6.0
tqdm>=4.67
notebook
jupyterlab